+++ lesson5.2/src/Main.java
@@ -1,4 +1,5 @@
 import java.util.Scanner;
+
 /*
         Задано перечисление Seasons, которое описывает четыре времени года.
         Реализовать ввод с клавиатуры номера времени года (1..4). На основе номера времени года вычислить продолжительность времени года в днях.
