+++ lesson5.2/src/MainClothes.java
@@ -9,11 +9,11 @@ import clothes.*;
       Создать классы наследники Одежды - Футболка (реализует интерфей сы "Мужская Одежда" и "Женская Одежда"), Штаны (реализует интерфейсы "Мужская Одежда" и "Женская Одежда"), Юбка (реализует интерфейсы "Женская Одежда"), Галстук (реализует интерфейсы "Мужская Одежда").
       Создать массив, содержащий все типы одежды. Создать класс Ателье, содержащий методы одетьЖенщину, одетьМужчину, на вход которых будет поступать массив, содержащий все типы одежды. Метод одетьЖенщину выводит на консоль всю информацию о женской одежде. То же самое для метода одетьМужчину.
       */
-public class Main_clothes {
+public class MainClothes {
     public static void main(String[] args) {
         SizeEnum coatSize = SizeEnum.XXS;
 
-        coatSize.getDescription();
+        System.out.println(coatSize.getDescription());
         int value = coatSize.getEuroSize();
         System.out.println(value);
