+++ lesson5.2/src/SizeEnum.java
@@ -1,0 +1,29 @@
+public enum SizeEnum {
+    XXS(32) {
+        @Override
+        public String getDescription() {
+            return "Детский размер";
+        }
+    },
+    XS(34),
+    S(36),
+    M(38),
+    L(40);
+
+
+    private int euroSize;
+
+    SizeEnum(int euroSize) {
+        this.euroSize = euroSize;
+    }
+
+    public int getEuroSize() {
+        return this.euroSize;
+    }
+
+    public String getDescription() {
+        return "Взрослый размер";
+    }
+}
+
+
