+++ lesson5.2/src/clothes/Skirt.java
@@ -1,6 +1,6 @@
 package clothes;
 
-public class Skirt extends Clothe implements Womenwear {
+public class Skirt extends Clothe implements WomenWear {
 
     public Skirt(int sizewear, int costwear, String colorwear) {
         super(sizewear, costwear, colorwear);
